+++ R/explain.r
@@ -13,6 +13,7 @@
 #' @export
 #' @param x An object to explain
 #' @param ... Other parameters possibly used by generic
+#' @return The first argument, invisibly.
 #' @examples
 #' \donttest{
 #' if (require("RSQLite") && has_lahman("sqlite")) {
@@ -49,7 +50,7 @@ explain.tbl_sql <- function(x, ...) {
   message("\n")
   message("<PLAN>\n", db_explain(con, sql_render(x, con = con)))
 
-  invisible(NULL)
+  invisible(x)
 }
 
 #' @export
@@ -59,4 +60,6 @@ show_query <- function(x) {
   on.exit(con_release(x$src, con), add = TRUE)
 
   message("<SQL>\n", sql_render(x, con = con))
+
+  invisible(x)
 }